# Performance improvement of `SparsePauliOp.compose`